исправлены и доработаны ошибки кроме utils.js по последнему ревью
